Create structStudent
Program which allows the user to input a number of students in an university. Than randomly generates a dynamic array of students, which are represented with structs which contain the students index number, year of study and average grade. The array of students is filled and printed using recursive functions.  At the end a function prints some statistics about the students  such as how many are failing, how many have excellent grades and how many students are in each year..